Make ASan report file and line number
Strangely enough, while a single compile-and-link command generates exe.dSYM/
dir that contains debug infos, separate compile/link invocation does not,
disabling file and line numbers report in ASan. This issue was directly
addressed in [1]. The reason for this is explained in [2].

This adds a post-build rule that runs 'dsymutil' which produces the debug
info dir from the executable.

This issue has been bugging me for a whole day now. Hopes this fixes any
debug symbol related problem once and for all.

[1] https://github.com/google/sanitizers/issues/207
[2] https://stackoverflow.com/questions/32297349/why-does-a-2-stage-command-line-build-with-clang-not-generate-a-dsym-directory