Add basic testing infrastructure
'ninja check' builds and runs Catch test executable defined in a separate
source. Currently it is a bogus test that always fails. The enormous increase
in line number is attributed to the Catch header.

Had to do some workaround for CTest, which somehow ignores its dependency on
the test target. See comment in CMakeLists.txt.

Also added lexing capability for parens and curly brackets.