Move skipWhitespace() for simpler EOF detection
OK, so putting skipWhitespace() either at the beginning or end of Lex()
is merely a problem of aesthetics. Putting it at the beginning takes into
account files starting with whitespaces, but requires additional skip at
the end of the file to properly detect EOF. Putting it at the end is
just the opposite: it eliminates this additional skip at the EOF, but
then again needs one at the start.

I personally prefer the latter as it always neatly prepares a valid
token character at the current look, and we have initialization method
New() where we can hide in the initial skip. This is the choice LBC!
makes as well.