Fix wrong column number calculation
...and write a basic pretty print function of tokens, for debugging
purpose.

Column numbers were off by one (and line numbers as well for EOLs),
because the offset of each BOL were wrongly set as that of the _start_
of the leading EOL, not the _end_.

Reported column numbers are left to start from 0, as the OCaml
compiler seems to do so as well.