Use CRTP to enable calling default visitors from derived...
...visitors.

The goal is to provide default visitor implementations, that do nothing
but just walk the subnodes, to the derived visitor class.  This way the
derived class doesn't have to define all of the visitors for every types
of node.  For this to be accomplished, we need a way to 'descend' into
the derived visitor method from a base visitor, and vice versa.

One way of achieving this is to declare all of the visitor methods as
virtual, which introduces a dynamic dispatch.  It turns out that the
dynamic dispatch is entirely unnecessary: from the user code, visitor
methods are supposed to be called upon a derived visitor class whose
type is statically known, e.g. 'NameBinder::visit()'.  This means that
it is already known at the compile time which of the overriden versions
of visitXXX() methods should be called.

A popular way to achieve this kind of static polymorphism in C++ is the
curiously recurring template pattern (CRTP).  It essentially encodes the
type of the derived class into the base class, making it possible for
the base to cast itself ('this' pointer) to the derived type and calling
its method.  We define a convenience function called 'dis()' that does
the cast, and convert all of the visitor and walker invocations to be
done through the cast pointer.